Updated with a new model -- failed
There are a couple of times the vehicle slows down at the red and then
resumes its top speed. One time it stops at red. Guess this results
from the fact that I set the image size to 256X256, which is hard for
the classification of traffic lights as they are tiny after shrinking.